+++ R/googleAuthR_auth.R
@@ -6,8 +6,7 @@ Authentication <- R6::R6Class(
   "Authentication",
   public = list(
     token = NULL,
-    shiny = FALSE,
-    app_url = NULL
+    method = NULL
   ),
   lock_objects = F,
   parent_env = emptyenv()
@@ -102,22 +101,56 @@ gar_auth <- function(token = NULL,
     stopifnot(is_legit_token(google_token))
     
     Authentication$set("public", "token", google_token, overwrite=TRUE)
+    Authentication$set("public", "method", "new_token", overwrite=TRUE)
     
   } else {
     ## supplied a file path or Token object
     if(is_legit_token(token)) {
+      ## a token object, just return it back
       google_token <- token
+      Authentication$set("public", "method", "passed_token", overwrite=TRUE)
     } else {
-      myMessage("Reading token from file path", level = 2)
-      google_token <- read_cache_token(token_path = getOption("googleAuthR.httr_oauth_cache"))
+      ## a file path to a token
+      options("googleAuthR.httr_oauth_cache" = token)
+      google_token <- read_cache_token(token_path = token)
+      google_token$cache_path <- token
+      Authentication$set("public", "method", "filepath", overwrite=TRUE)
     }
     
     Authentication$set("public", "token", google_token, overwrite=TRUE)
     
   }
-  
+  gar_token_info()
   return(invisible(Authentication$public_fields$token)) 
   
+}
+
+#' Get current token summary
+#' 
+#' Get details on the current active auth token to help debug issues
+#' 
+#' @export
+gar_token_info <- function(){
+  token <- Authentication$public_fields$token
+  method <- Authentication$public_fields$method
+  
+  myMessage("Token cache file: ", token$cache_path, level = 3)
+  myMessage("Scopes: ", paste(token$params$scope, collapse = " "), level = 2)
+  myMessage("Hash: ", token$hash(), level = 2)
+  
+  if(!is.null(token$app$key)){
+    myMessage("App key: ", token$app$key, level = 2)
+  }
+
+  myMessage("Method: ", method, level = 2)
+  
+  ## service
+  if(!is.null(token$secrets)){
+    myMessage("Type: ", token$secrets$type, level = 2)
+    myMessage("ProjectID: ", token$secrets$project_id, level = 2)
+    myMessage("Client email: ", token$secrets$client_email, level = 2)
+    myMessage("ClientID: ", token$secrets$client_id, level = 2)
+  }
 }
 
 ## httr cache files such as .httr-oauth can hold multiple tokens for different scopes.
@@ -132,6 +165,12 @@ read_cache_token <- function(token_path = getOption("googleAuthR.httr_oauth_cach
     stop(sprintf("Cannot read token from alleged .rds file:\n%s",
                  token_path))
   }
+  
+  if(inherits(google_token, "list")){
+    myMessage("Found a list, return token in first element.")
+    google_token <- google_token[[1]]
+  }
+  
   google_token
 }
 
@@ -155,7 +194,7 @@ get_google_token <- function(shiny_return_token=NULL) {
     
     
   } else { #shiny session
-    
+    Authentication$set("public", "method", "shiny", overwrite=TRUE)
     token <- shiny_return_token
     
   }
@@ -216,6 +255,12 @@ is_legit_token <- function(x) {
   
   if(!inherits(x, "Token2.0")) {
     myMessage("Not a Token2.0 object. Found:", class(x), level=2)
+    if(!inherits(x, "list")){
+      if(inherits(x[[1]], "Token2.0")){
+        myMessage("Its a list of Token2.0 objects though")
+        return(TRUE)
+      }
+    }
     return(FALSE)
   }
   
@@ -284,7 +329,7 @@ gar_auth_service <- function(json_file, scope = getOption("googleAuthR.scopes.se
   google_token <- httr::oauth_service_token(endpoint, secrets, scope)
   
   Authentication$set("public", "token", google_token, overwrite=TRUE)
-  
+  Authentication$set("public", "method", "service_json", overwrite=TRUE)
   myMessage("Returning service token", level=1)
   
   return(invisible(Authentication$public_fields$token))
@@ -353,10 +398,11 @@ gar_gce_auth <- function(service_account = "default",
   
   ## for other google auth on a server (such as Google Analytics) need to manually do tokens via OOB
   options(httr_oob_default = TRUE)
+  Authentication$set("public", "method", "gce_auth", overwrite=TRUE)
   
   ## puts it in environment
   gar_auth(token_formatted)
-  
+
 }