+++ tests/testthat/test-gaAuthR.R
@@ -54,8 +54,7 @@ test_that("The auth httr file can be found",{
 
 test_that("Can authenticate .httr passed as a file", {
   skip_on_cran()
-  options(googleAuthR.httr_oauth_cache = "httr-oauth.rds")
-  expect_s3_class(gar_auth("httr-oauth.rds")[[1]], "Token2.0")
+  expect_s3_class(gar_auth("httr-oauth.rds"), "Token2.0")
   
 })